pw: loop: Make `Loop` trait no longer be unsafe, mark `as_ptr` as unsafe instead.
`unsafe` traits are better suited where implementing the trait at all can lead to unsafety, e.g. the `Send` trait.

The only part unsafe about the loop trait is the `as_ptr` method which must point to a valid pw_loop.

Clippy also wants unsafe traits to have a `Safety` section, but that section is more relevant directly on the method, and is also added by this commit.